Add sidesway inhibited effective length factor
Also retool the internals of the effective length factor calculation
to use the brentq solver instead of fsolve, since it only needs to find
a single scalar root. This also removes the need (and ability) to provide
an initial guess for K. Instead, a bracket is provided which is already
established by the AISC nomograms.